Proof of concept for handling two concurrent asynchronous tasks (UI & background)
Task 1 simulates a long-running UI operation such as launching a new fragment (4 seconds)
Task 2 simulates a long-running restartable back-end operation that returns a result (8 seconds)

Task 1 should not be interrupted by additional user requests (because creating a UI widget such as a fragment can't be canceled).  However Task 2 can be canceled.  Therefore Task 1 is a SideEffect whereas Task 2 is a CancellableSideEffect.

Sequence of events:
1. Activity code registers observers for changes to isShowingResult1 and isShowingResult2
2. User taps button
3. Two async tasks triggered
   - Task 1 simulates a long-running UI operation such as launching a new fragment (4 seconds)
     a. Update the state's result1 to "[showUIAsync]"
     b. A SideEffect initializes a flag (isShowingResult1) to true
        b1. Activity's observer sees that isShowingResult1 was changed to true
        b2. Disable the button to prevent multiple requests (suggested by https://medium.com/androiddevelopers/coroutines-on-android-part-iii-real-work-2ba8a2ec2f45)
        b3. Activity shows a TextField and delays 4 seconds (as a simulation of time needed to construct an elaborate UI such as a fragment)
            b3a. The TextField shows "[showUIAsync]"
        b4. After the delay, Activity dispatches action saying that the UI is rendered
        b5. Activity's observer sees that isShowingResult1 was changed to false
        b6. Re-enable the button (thus allowing Task 2 to be canceled and restarted if user wants)
   - Task 2 simulates a long-running back-end operation that returns a result (8 seconds)
     a. Update state's result2 to "[startBackendAsync]"
     b. A CancellableSideEffect launches a long-running task (8 seconds)
        b1. If there user taps the button again at this point, the side effect gets canceled and it restarts on the previous step
     c. After the 8 seconds, it dispatches an action to indicate that it's finished
     d. A SideEffect initializes a flag (isShowingResult2) to true
        d1. Activity's observer sees that isShowingResult2 was changed (the value doesn't matter)
        d2. Activity shows a TextField
            b2a. The TextField shows "[startBackendAsync, endBackendAsync]"

In addition to all the steps above, the system also updates an Espresso Idling Resource counter so that the test framework knows to wait for both async tasks to finish.  When the user taps the button, the counter is incremented 2 because there are two idling resources (i.e. the two tasks are resources that are "idle" while they wait to complete).  When each one complete is decrements.  So when they both complete, we reach 0 and the test framework knows it can continue (in this case to perform our validation logic).

Testing:
-shouldPerformTwoConcurrentAsyncTasks